breaking: Transport.GetMaxPacketSize changed to ushort. Limits us to 64KB max size which realistically is almost too much for multiplayer games. This will make pooling significantly easier because we can pool all readers/writers with ushort.max size and don't need resizing/buffer swapping anymore.

+++ Assets/Mirror/Runtime/Transport/KCP/MirrorTransport/KcpTransport.cs
@@ -120,7 +120,7 @@ namespace kcp2k
         public override void Shutdown() {}
 
         // MTU
-        public override int GetMaxPacketSize(int channelId = Channels.DefaultReliable) => Kcp.MTU_DEF;
+        public override ushort GetMaxPacketSize(int channelId = Channels.DefaultReliable) => Kcp.MTU_DEF;
 
         public override string ToString()
         {
