+++ Assets/Mirror/Runtime/Transport/MiddlewareTransport.cs
@@ -27,7 +27,7 @@ namespace Mirror
         }
 
         public override bool Available() => inner.Available();
-        public override int GetMaxPacketSize(int channelId = 0) => inner.GetMaxPacketSize(channelId);
+        public override ushort GetMaxPacketSize(int channelId = 0) => inner.GetMaxPacketSize(channelId);
         public override void Shutdown() => inner.Shutdown();
 
         #region Client
