+++ Assets/Mirror/Tests/Editor/MiddlewareTransportTest.cs
@@ -46,10 +46,10 @@ namespace Mirror.Tests
         }
 
         [Test]
-        [TestCase(Channels.DefaultReliable, 4000)]
-        [TestCase(Channels.DefaultReliable, 2000)]
-        [TestCase(Channels.DefaultUnreliable, 4000)]
-        public void TestGetMaxPacketSize(int channel, int packageSize)
+        [TestCase(Channels.DefaultReliable, (ushort)4000)]
+        [TestCase(Channels.DefaultReliable, (ushort)2000)]
+        [TestCase(Channels.DefaultUnreliable, (ushort)4000)]
+        public void TestGetMaxPacketSize(int channel, ushort packageSize)
         {
             inner.GetMaxPacketSize(Arg.Any<int>()).Returns(packageSize);
 
@@ -159,7 +159,7 @@ namespace Mirror.Tests
             middleware.ServerSend(id, channel, segment);
 
             inner.Received(1).ServerSend(id, channel, Arg.Is<ArraySegment<byte>>(x => x.Array == array && x.Offset == offset && x.Count == count));
-            // only need to check first arg, 
+            // only need to check first arg,
             inner.Received(0).ServerSend(Arg.Is<int>(x => x != id), Arg.Any<int>(), Arg.Any<ArraySegment<byte>>());
         }
